+++ droidar/src/util/ImageTransform.java
@@ -73,8 +73,8 @@ public class ImageTransform {
 			float newWidth) {
 		int width = bitmap.getWidth();
 		int height = bitmap.getHeight();
-		float scaleWidth = ((float) newWidth) / width;
-		float scaleHeight = ((float) newHeight) / height;
+		float scaleWidth = newWidth / width;
+		float scaleHeight = newHeight / height;
 		Matrix matrix = new Matrix();
 		matrix.postScale(scaleWidth, scaleHeight);
 		return Bitmap.createBitmap(bitmap, 0, 0, width, height, matrix, true);
